changed the number of part and note back to byte from ushort

+++ Dissimilis.WebAPI/DTOs/NewNoteDTO.cs
@@ -8,7 +8,7 @@ namespace Dissimilis.WebAPI.DTOs
     public class NewNoteDTO
     {
         public int BarId { get; set; }
-        public ushort NoteNumber { get; set; }
+        public byte NoteNumber { get; set; }
         public byte Length { get; set; }
         public string[] Notes { get; set; }
