+++ Dissimilis.WebAPI/DTOs/NewPartDTO.cs
@@ -10,7 +10,7 @@ namespace Dissimilis.WebAPI.DTOs
     {
         public int SongId { get; set; }
         public string Title { get; set; }
-        public ushort PartNumber { get; set; }
+        public byte PartNumber { get; set; }
         public NewBarDTO[] Bars { get; set; }
 
         public NewPartDTO() { }
